pyproject.toml
[tool.poetry]
name = "mochi-code"
version = "0.1.0"
description = "An AI pair programmer"
authors = ["Diogo Neves <diogo.neves@gmail.com>"]
readme = "README.md"
packages = [{ include = "mochi_code" }]

[tool.poetry.scripts]
mochi = "mochi_code.mochi:cli"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.181"
python-dotenv = "^1.0.0"
openai = "^0.27.7"
pydantic = "1.10.8"
retry2 = "^0.9.5"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pylint = "^2.17.4"
mypy = "^1.3.0"
yapf = "^0.33.0"
types-retry = "^0.9.9.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
